Move UI and Parser to respective files
