+++ data/userData.csv
@@ -1,4 +1,1 @@
-Deadline,bio homework,true,2025-11-25
-Deadline,chem hw,false,2025-12-24
-Event,party at the hollywood hills,false,2025-08-15,2025-09-10
-Todo,code,false
+Todo,game,false
